replace obsolete LifecycleActivity https://medium.com/@mladenrakonjac/note-since-the-architecture-components-are-in-alpha-stage-fragment-and-appcompatactivity-classes-b4b055ecd35b
"Note: Since the Architecture Components are in alpha stage, Fragment and AppCompatActivity classes cannot implement it (because we cannot add a dependency from a stable component to an unstable API). Until Lifecycle is stable, LifecycleActivity andLifecycleFragment classes are provided for convenience. After the Lifecycles project is released, support library fragments and activities will implement the LifecycleOwner interface; LifecycleActivity and LifecycleFragment will be deprecated at that time. Also, see Implementing LifecycleOwner in custom activities and fragments."